Option.run() can be overriden by an anonymous class that extends Option, to then run some LOCAL (i.e. in the class that instantiates it) method.
To see how to do it, take a gander at InputPrompterTest. In that file there are concrete examples about how to create the options that can be run as a function.

Also, another cool thing I did was check that the print statements are printing an exact string. This is important for phase 1 because we can then know for sure that a menu will come out the way we want it without having to manually test.