Refactored MenuController and UserController to implement InputController
Option is now an adaptor for Permission and Template. I'm not sure if this is a good change or not, but it may very well be worth it.

Fixed test cases. Turns out that you need to set the input at the top, before the scanner is instantiated in helper functions?